Move Navbar-components into Navbar-folder
