+++ src/lib/components/Menu.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { toolState, ToolState } from '$lib/stores/stateStore';
 
-	import MenuButton from './MenuButton.svelte';
+	import MenuButton from './Navbar/MenuButton.svelte';
 	import Pan from '$lib/components/tools/Pan.svelte';
 	import Move from '$lib/components/tools/Move.svelte';
 	import Draw from '$lib/components/tools/Draw.svelte';
