+++ src/lib/components/Navbar/Logo.svelte
