+++ src/lib/components/UI.svelte
@@ -3,7 +3,7 @@
 	import ServerId from './ServerId.svelte';
 	import Menu from './Menu.svelte';
 	import ChatButton from './ChatButton.svelte';
-	import Navbar from './Navbar.svelte';
+	import Navbar from './Navbar/Navbar.svelte';
 	import Chat from './Chat.svelte';
 
 	import { viewChat } from '$lib/stores/stateStore';
