+++ src/lib/components/tools/Zoom.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import MenuButton from '$lib/components/MenuButton.svelte';
+	import MenuButton from '$lib/components/Navbar/MenuButton.svelte';
 	import Icons from '$lib/icons/MenuIcons.json';
 	import { canvasView } from '$lib/stores/stateStore';
